fix(components/icon): reorganize internal file structure for correct documentation output
