+++ libs/components/icon/src/index.ts
@@ -1,7 +1,7 @@
-export { SkyIconStackItem } from './lib/icon/icon-stack-item';
-export { SkyIconModule } from './lib/icon/icon.module';
-export { SkyIconType } from './lib/icon/types/icon-type';
-export { SkyIconVariantType } from './lib/icon/types/icon-variant-type';
+export { SkyIconStackItem } from './lib/modules/icon/icon-stack-item';
+export { SkyIconModule } from './lib/modules/icon/icon.module';
+export { SkyIconType } from './lib/modules/icon/types/icon-type';
+export { SkyIconVariantType } from './lib/modules/icon/types/icon-variant-type';
 
-export { SkyIconComponent as λ1 } from './lib/icon/icon.component';
-export { SkyIconStackComponent as λ2 } from './lib/icon/icon-stack.component';
+export { SkyIconComponent as λ1 } from './lib/modules/icon/icon.component';
+export { SkyIconStackComponent as λ2 } from './lib/modules/icon/icon-stack.component';
