+++ libs/components/icon/src/lib/modules/icon/fixtures/icon-fixtures.module.ts
