+++ libs/components/icon/src/lib/modules/icon/fixtures/icon-stack.component.fixture.ts
