+++ libs/components/icon/src/lib/modules/icon/fixtures/icon.component.fixture.html
