+++ libs/components/icon/src/lib/modules/icon/icon-class-list.pipe.ts
