+++ libs/components/icon/src/lib/modules/icon/icon-resolver.service.ts
