+++ libs/components/icon/src/lib/modules/icon/icon-stack-item.ts
