+++ libs/components/icon/src/lib/modules/icon/icon-stack.component.scss
