+++ libs/components/icon/src/lib/modules/icon/icon-svg-resolver.service.ts
