+++ libs/components/icon/src/lib/modules/icon/icon-svg.component.ts
