+++ libs/components/icon/src/lib/modules/icon/icon-svg.default.component.scss
