+++ libs/components/icon/src/lib/modules/icon/icon-svg.modern.component.scss
