+++ libs/components/icon/src/lib/modules/icon/icon-variant.ts
