+++ libs/components/icon/src/lib/modules/icon/icon.component.spec.ts
