+++ libs/components/icon/src/lib/modules/icon/icon.module.ts
