+++ libs/components/icon/src/lib/modules/icon/types/icon-resolved.ts
