+++ libs/components/icon/src/lib/modules/icon/types/icon-variant-type.ts
