+++ libs/components/icon/testing/src/public-api.ts
